Fixes missing dependency exception under minimal install for brotli optional extra.
Moves compression config enums into config.compression module.
